requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.1.0=pypi_0
argon2-cffi=20.1.0=py38h2bbff1b_1
astor=0.8.1=pypi_0
async_generator=1.10=pyhd3eb1b0_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
base58=2.1.0=pypi_0
beautifulsoup4=4.9.3=pyhb0f4dca_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blinker=1.4=pypi_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2021.5.25=haa95532_1
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py38haa95532_0
cffi=1.14.5=py38hcd4344a_0
chardet=4.0.0=py38haa95532_1003
click=7.1.2=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
console_shortcut=0.1.1=4
cryptography=3.4.7=py38h71e12ea_0
cycler=0.10.0=py38_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py38_0
et_xmlfile=1.1.0=py38haa95532_0
freetype=2.10.4=hd328e21_0
gitdb=4.0.7=pypi_0
gitpython=3.1.18=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.7.3=py38haa95532_1
importlib_metadata=3.7.3=hd3eb1b0_1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.21.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
jdcal=1.4.1=py_0
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.1=py38haa95532_0
jupyterlab-widgets=1.0.0=pypi_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=py38hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.2.0=hd0e1b90_0
lz4-c=1.9.3=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38he774522_0
matplotlib=3.3.4=py38haa95532_0
matplotlib-base=3.3.4=py38h49ac443_0
mistune=0.8.4=py38he774522_1000
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.3.0=py38h46781fe_0
mkl_random=1.1.1=py38h47e9c7a_0
msys2-conda-epoch=20160418=1
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.2=pyhd3eb1b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.3.0=py38haa95532_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
olefile=0.46=py_0
openpyxl=3.0.7=pyhd3eb1b0_0
openssl=1.1.1k=h2bbff1b_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.3=py38hf11a4ad_0
pandoc=2.12=haa95532_0
pandocfilters=1.4.3=py38haa95532_1
parso=0.8.1=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py38h4fa10fc_0
pip=21.0.1=py38haa95532_0
plotly=4.14.3=pyhd3eb1b0_0
powershell_shortcut=0.0.1=3
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.17.3=pypi_0
pyarrow=4.0.1=pypi_0
pycparser=2.20=py_2
pydeck=0.6.2=pypi_0
pygments=2.8.1=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.17.3=py38he774522_0
pysocks=1.7.1=py38haa95532_0
python=3.8.8=hdbf39b2_4
python-dateutil=2.8.1=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pywin32=227=py38he774522_1
pywinpty=0.5.7=py38_0
pyzmq=20.0.0=py38hd77b12b_1
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pyhd3eb1b0_0
retrying=1.3.3=py_2
scikit-learn=0.24.1=py38hf11a4ad_0
scipy=1.6.2=py38h14eb087_0
seaborn=0.11.1=pyhd3eb1b0_0
selenium=3.141.0=py38h2bbff1b_1000
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
smmap=4.0.0=pypi_0
soupsieve=2.0.1=py_0
sqlite=3.35.2=h2bbff1b_0
streamlit=0.83.0=pypi_0
terminado=0.9.3=py38haa95532_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
toml=0.10.2=pypi_0
toolz=0.11.1=pypi_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.0.5=pyhd3eb1b0_0
tzlocal=2.1=pypi_0
urllib3=1.26.4=pyhd3eb1b0_0
validators=0.18.2=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.2=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
winpty=0.4.3=4
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h62dcd97_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
